requirements.txt
attrs
urwid
pytest
PyYAML
